error_exception_util: add a function template overloaded Require(); also rectify parameter type of an old Require().
The function template is added because we need an approach to throw a specific exception, which we know how to handle it, when the condition is violated. RAISE and RAISE_WITH_DUMP represent agnostic exceptions that we know little about, and have no way to handle/recover them in place.

Also be noted that, the type const std::string of the parameter in an old Require() function is definitely a typo.