Add macro RETAIN_LIFETIME_TO_HERE()
Sometimes we devide definition and assignment of a variable on purpose
because the assignment is conditional but the lifetime of the variable
should never be restricted within the condition-block.

However, code analyzer or linter may suggest join the definition and the
assignment; with this macro, we can get rid of this tiresome hint, and
meanwhile, make our intention more obvious.